+++ backend/src/routes/complaints.routes.ts
@@ -1,8 +1,8 @@
 import express from 'express'
-import * as controller from '../controllers/complaints.controller'
+import * as controller from '../controllers/complaint.controller'
 
 const router = express.Router()
-
+/*
 router.get('/', controller.getAllReviews)
-
+*/
 export default router
